Finished

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,14 @@ import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        authenticatedUserId
+        const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
         res.status(200).json(notes);
     } catch (error) {
         //next(error);
@@ -14,8 +18,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -24,6 +30,9 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, " You cannot access this note")
         }
 
         res.status(200).json(note);
@@ -40,13 +49,16 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Note must have a title");
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -66,12 +78,14 @@ interface UpdateNoteBody {
     text?: string,
 }
 
-export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async(req, res, next) => {
+export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -86,6 +100,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, " You cannot access this note")
+        }
+
         note.title = newTitle;
         note.text = newText;
 
@@ -97,10 +115,12 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     }
 };
 
-export const deleteNote: RequestHandler = async(req, res, next) => {
+export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -110,11 +130,13 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
         if (!note) {
             throw createHttpError(404, "Note not found");
         }
-        // nie działa
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, " You cannot access this note")
+        }
         await note.deleteOne()
-        
+
         res.sendStatus(204);
     } catch (error) {
-      next(error);  
+        next(error);
     }
 };
